Move subset of messages logic into a hook
