+++ apps/tlon-web/src/chat/ChatWindow.tsx
@@ -1,4 +1,3 @@
-import { PostTuple } from '@tloncorp/shared/dist/urbit/channel';
 import bigInt from 'big-integer';
 import React, {
   ReactElement,
@@ -6,7 +5,6 @@ import React, {
   useEffect,
   useMemo,
   useRef,
-  useState,
 } from 'react';
 import { useNavigate, useParams, useSearchParams } from 'react-router-dom';
 import { VirtuosoHandle } from 'react-virtuoso';
@@ -16,13 +14,11 @@ import ChatUnreadAlerts from '@/chat/ChatUnreadAlerts';
 import EmptyPlaceholder from '@/components/EmptyPlaceholder';
 import ArrowS16Icon from '@/components/icons/ArrowS16Icon';
 import { useChannelCompatibility } from '@/logic/channel';
-import { log } from '@/logic/utils';
+import useSubsetOfMessagesForScroller from '@/logic/useSubsetOfMessagesForScroller';
 import { useInfinitePosts, useMarkReadMutation } from '@/state/channel/channel';
 
 import ChatScrollerPlaceholder from './ChatScroller/ChatScrollerPlaceholder';
 import { useChatInfo, useChatStore } from './useChatStore';
-
-const WINDOW_SIZE = 30;
 
 interface ChatWindowProps {
   whom: string;
@@ -60,22 +56,6 @@ export default function ChatWindow({
     isFetchingPreviousPage,
   } = useInfinitePosts(nest, scrollToId);
   const { mutate: markRead } = useMarkReadMutation();
-  const totalMessagesCached = messages ? messages.length : 0;
-  const [scrollerMessages, setScrollerMessages] = useState<PostTuple[]>([]);
-  const [shouldInvert, setShouldInvert] = useState(true);
-  const [scrollerMessagesSliceStart, setScrollerMessagesSliceStart] =
-    useState<number>();
-  const [scrollerMessagesSliceEnd, setScrollerMessagesSliceEnd] =
-    useState<number>();
-  const [currentPageIndex, setCurrentPageIndex] = useState(1);
-  const [fakeLoadingOlder, setFakeLoadingOlder] = useState(false);
-  const [fakeLoadingNewer, setFakeLoadingNewer] = useState(false);
-  const totalPagesCached = useMemo(
-    () => Math.ceil(totalMessagesCached / WINDOW_SIZE),
-    [totalMessagesCached]
-  );
-  const [initialLoad, setInitialLoad] = useState(true);
-  const seenScrollToIdRef = useRef(false);
   const scrollerRef = useRef<VirtuosoHandle>(null);
   const readTimeout = useChatInfo(whom).unread?.readTimeout;
   const clearOnNavRef = useRef({ readTimeout, nest, whom, markRead });
@@ -89,20 +69,37 @@ export default function ChatWindow({
         : latestMessageIndex,
     [scrollToId, messages, latestMessageIndex]
   );
+  const {
+    scrollerMessages,
+    onAtBottom,
+    onAtTop,
+    shouldInvert,
+    hasLoadedNewest,
+    hasLoadedOldest,
+    isLoadingNewer,
+    isLoadingOlder,
+  } = useSubsetOfMessagesForScroller(
+    messages,
+    isLoading,
+    isFetching,
+    hasPreviousPage,
+    isFetchingPreviousPage,
+    hasNextPage,
+    fetchPreviousPage,
+    fetchNextPage,
+    isFetchingNextPage,
+    whom,
+    nest,
+    scrollToId,
+    scrollToIndex,
+    markRead
+  );
   const msgIdTimeInMessages = useMemo(
     () =>
       scrollToId
         ? messages.findIndex((m) => m[0].toString() === scrollToId) !== -1
         : false,
     [scrollToId, messages]
-  );
-  const msgIdTimeInScrollerMessages = useMemo(
-    () =>
-      scrollToId
-        ? scrollerMessages.findIndex((m) => m[0].toString() === scrollToId) !==
-          -1
-        : false,
-    [scrollToId, scrollerMessages]
   );
   const latestIsMoreThan30NewerThanScrollTo = useMemo(
     () =>
@@ -139,241 +136,12 @@ export default function ChatWindow({
     root,
   ]);
 
-  const lastScrollerMessagesSliceStart = useRef<number | undefined>(undefined);
-  const lastScrollerMessagesSliceEnd = useRef<number | undefined>(undefined);
-  const lastWhom = useRef<string | undefined>(undefined);
-
-  useEffect(() => {
-    // scrollerMessages only updates when slice start and end change
-    // or on initial load
-
-    if (initialLoad && totalMessagesCached > 0 && !isLoading && !scrollToId) {
-      // if we're in the initial load, and we have messages, and we're not
-      // fetching, and we don't have a scrollToId, then we should set the
-      // scrollerMessages to the default slice start/end
-
-      setScrollerMessagesSliceStart(totalMessagesCached - WINDOW_SIZE);
-      setScrollerMessagesSliceEnd(totalMessagesCached);
-      setCurrentPageIndex(1);
-
-      setScrollerMessages(
-        messages.slice(scrollerMessagesSliceStart, scrollerMessagesSliceEnd)
-      );
-
-      setInitialLoad(false);
-    }
-
-    if (
-      !initialLoad &&
-      !fakeLoadingNewer &&
-      !fakeLoadingOlder &&
-      totalMessagesCached > 0 &&
-      scrollerMessagesSliceStart !== undefined &&
-      scrollerMessagesSliceStart >= 0 &&
-      scrollerMessagesSliceEnd !== undefined &&
-      scrollerMessagesSliceEnd <= totalMessagesCached &&
-      lastWhom.current === whom &&
-      (lastScrollerMessagesSliceStart.current !== scrollerMessagesSliceStart ||
-        lastScrollerMessagesSliceEnd.current !== scrollerMessagesSliceEnd)
-    ) {
-      // if we're not in the initial load, and we're not faking loading newer
-      // or older messages, and we have a slice start and end, and the slice
-      // start and end are within the bounds of the total messages cached, and
-      // the whom hasn't changed, and the slice start and end have changed, then
-      // we should update the scrollerMessages
-
-      setScrollerMessages(
-        messages.slice(scrollerMessagesSliceStart, scrollerMessagesSliceEnd)
-      );
-      lastScrollerMessagesSliceStart.current = scrollerMessagesSliceStart;
-      lastScrollerMessagesSliceEnd.current = scrollerMessagesSliceEnd;
-    }
-  }, [
-    messages,
-    scrollerMessagesSliceStart,
-    scrollerMessagesSliceEnd,
-    fakeLoadingNewer,
-    fakeLoadingOlder,
-    totalMessagesCached,
-    whom,
-    initialLoad,
-    isLoading,
-    scrollToId,
-  ]);
-
   useEffect(() => {
     useChatStore.getState().setCurrent(whom);
     return () => {
       useChatStore.getState().setCurrent('');
     };
   }, [whom]);
-
-  useEffect(() => {
-    // channel switch
-
-    if (whom !== lastWhom.current && !scrollToId) {
-      lastWhom.current = whom;
-      setInitialLoad(true);
-    }
-  }, [whom, totalMessagesCached, scrollToId]);
-
-  useEffect(() => {
-    // if we have a scrollToId and we're in the initial load, and the scrollToId
-    // is not in the current set of scrollerMessages, then we need to update the
-    // slice start and end to include the scrollToId
-
-    if (initialLoad && scrollToId && !msgIdTimeInScrollerMessages) {
-      const sliceStart = scrollToIndex - WINDOW_SIZE / 2;
-      const sliceEnd = scrollToIndex + WINDOW_SIZE / 2;
-      const pageIndex =
-        1 +
-        totalPagesCached -
-        Math.ceil((scrollToIndex / totalMessagesCached) * totalPagesCached);
-
-      if (sliceStart > 0) {
-        setScrollerMessagesSliceStart(sliceStart);
-        setScrollerMessagesSliceEnd(sliceEnd);
-        setCurrentPageIndex(pageIndex);
-        lastWhom.current = whom;
-        seenScrollToIdRef.current = true;
-        setInitialLoad(false);
-        setShouldInvert(false);
-      }
-    }
-  }, [
-    initialLoad,
-    whom,
-    messages,
-    currentPageIndex,
-    totalMessagesCached,
-    scrollToId,
-    scrollToIndex,
-    msgIdTimeInScrollerMessages,
-    totalPagesCached,
-  ]);
-
-  const onAtBottom = useCallback(() => {
-    setShouldInvert(false);
-
-    if (
-      currentPageIndex === 1 &&
-      scrollerMessagesSliceEnd !== totalMessagesCached &&
-      !isFetching &&
-      !fakeLoadingNewer &&
-      !fakeLoadingOlder
-    ) {
-      setFakeLoadingNewer(true);
-      setScrollerMessagesSliceEnd(totalMessagesCached);
-    }
-
-    if (
-      currentPageIndex !== 1 &&
-      !isFetching &&
-      scrollerMessagesSliceEnd &&
-      !fakeLoadingNewer &&
-      !fakeLoadingOlder
-    ) {
-      // if we're not at the bottom of the first page, and we're not fetching,
-      // and we're not faking loading, and we have a slice end, then
-      // we should fake load newer messages
-
-      setFakeLoadingNewer(true);
-      const maybeSliceEnd = scrollerMessagesSliceEnd + WINDOW_SIZE;
-      const sliceEnd =
-        maybeSliceEnd > totalMessagesCached
-          ? totalMessagesCached
-          : maybeSliceEnd;
-
-      setScrollerMessagesSliceEnd(sliceEnd);
-
-      setCurrentPageIndex((prev) => prev - 1);
-    }
-
-    const { bottom, delayedRead } = useChatStore.getState();
-    bottom(true);
-    delayedRead(whom, () => markRead({ nest }));
-    if (currentPageIndex === 1 && hasPreviousPage && !isFetching) {
-      log('fetching previous page');
-      fetchPreviousPage();
-    }
-
-    setTimeout(() => {
-      setFakeLoadingNewer(false);
-    }, 100);
-  }, [
-    nest,
-    whom,
-    markRead,
-    fetchPreviousPage,
-    hasPreviousPage,
-    isFetching,
-    currentPageIndex,
-    fakeLoadingNewer,
-    fakeLoadingOlder,
-    scrollerMessagesSliceEnd,
-    totalMessagesCached,
-  ]);
-
-  const onAtTop = useCallback(() => {
-    setShouldInvert(true);
-
-    if (
-      currentPageIndex !== totalPagesCached &&
-      !isFetching &&
-      !fakeLoadingOlder &&
-      !fakeLoadingNewer &&
-      scrollerMessagesSliceEnd
-    ) {
-      // if we're not at the top of the first page, and we're not fetching,
-      // and we're not faking loading, and we have a slice end, then we should
-      // fake load older messages
-
-      setFakeLoadingOlder(true);
-      const newPageIndex = currentPageIndex + 1;
-      const sliceStart =
-        totalMessagesCached -
-        (newPageIndex / totalPagesCached) * totalMessagesCached;
-
-      setScrollerMessagesSliceStart(sliceStart);
-      setCurrentPageIndex((prev) => prev + 1);
-    }
-
-    if (currentPageIndex === totalPagesCached && hasNextPage && !isFetching) {
-      log('fetching next page');
-      fetchNextPage();
-    }
-
-    setTimeout(() => {
-      setFakeLoadingOlder(false);
-    }, 300);
-  }, [
-    fetchNextPage,
-    hasNextPage,
-    isFetching,
-    currentPageIndex,
-    totalPagesCached,
-    scrollerMessagesSliceEnd,
-    fakeLoadingOlder,
-    fakeLoadingNewer,
-    totalMessagesCached,
-  ]);
-
-  useEffect(() => {
-    // if we have loaded the newest messages then we're at the bottom
-    // then we should be inverted
-    if (
-      currentPageIndex === 1 &&
-      scrollerMessagesSliceEnd === totalMessagesCached &&
-      !hasPreviousPage
-    ) {
-      setShouldInvert(true);
-    }
-  }, [
-    currentPageIndex,
-    scrollerMessagesSliceEnd,
-    totalMessagesCached,
-    hasPreviousPage,
-  ]);
 
   // read the messages once navigated away
   useEffect(() => {
@@ -444,8 +212,8 @@ export default function ChatWindow({
           key={whom}
           messages={scrollerMessages}
           shouldInvert={shouldInvert}
-          isLoadingOlder={fakeLoadingOlder ?? isFetchingNextPage}
-          isLoadingNewer={fakeLoadingNewer ?? isFetchingPreviousPage}
+          isLoadingOlder={isLoadingOlder}
+          isLoadingNewer={isLoadingNewer}
           whom={whom}
           topLoadEndMarker={prefixedElement}
           scrollTo={scrollToId ? bigInt(scrollToId) : undefined}
@@ -454,14 +222,8 @@ export default function ChatWindow({
           onAtBottom={onAtBottom}
           scrollElementRef={scrollElementRef}
           isScrolling={isScrolling}
-          hasLoadedOldest={
-            currentPageIndex === totalPagesCached && !hasNextPage
-          }
-          hasLoadedNewest={
-            currentPageIndex === 1 &&
-            scrollerMessagesSliceEnd === totalMessagesCached &&
-            !hasPreviousPage
-          }
+          hasLoadedOldest={hasLoadedOldest}
+          hasLoadedNewest={hasLoadedNewest}
         />
       </div>
       {scrollToId &&
